issue/#171-refactor-file-structure * refactor file structure  and update imports
