+++ frontend/src/components/styledComponents/StyledDivider.tsx
@@ -1,6 +1,6 @@
 import { Divider } from "@mantine/core";
 import { useColorPalette } from "../../hooks/useColorPalette";
-import { colorPaletteConstants } from "../../utility/Constants";
+import { colorPaletteConstants } from "../../utils/Constants";
 
 export default function StyledDivider() {
   const colorPalette = useColorPalette();
