+++ frontend/src/utils/ColorPalettes.tsx
