+++ frontend/src/utils/UtilityFunctions.tsx
